pyproject.toml
[project]
name = "dvc-pytorch-lightning-mlops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=24.8.0",
    "coverage>=7.6.1",
    "dvc>=3.55.2",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "lightning>=2.4.0",
    "loguru>=0.7.2",
    "matplotlib>=3.9.2",
    "numpy==1.26.0",
    "pandas>=2.2.3",
    "pytest-cov>=5.0.0",
    "pytest>=8.3.3",
    "rich>=13.9.2",
    "rootutils>=1.0.7",
    "seaborn>=0.13.2",
    "tensorboard>=2.18.0",
    "timm>=1.0.9",
    "torch==2.2.2",
    "torchmetrics>=1.4.2",
    "torchvision>=0.17.2",
    "typing-extensions>=4.12.2",
    "dvc-gdrive==3.0.1",
    "scikit-learn>=1.5.2",
    "dvc-s3==3.2.0",
    "boto3>=1.35.36",
    "split-folders>=0.5.1",
]
